Restructure source tree into feature-based folders
src/
  lib/firebase.js          — Firebase init (was buried in components/landingComp)
  components/
    landing/  match/  chat/  profile/  signup/  — renamed from *Comp
    common/SpinLoad.jsx    — shared loader (was loose in src/)
  store/
    store.js
    reducers/              — moved under the store they feed
  api/  matching/  hooks/  contexts/  data/  pages/  — unchanged

Removed dead files: server/algo.js (superseded by server/matching/),
pages/signedIn.jsx (unused barrel), pages/loader.jsx + css (unreferenced).

All import paths rewritten. Verified: eslint clean, matching test
suites pass, and every page (landing/login/about/signup + signed-in
profile/match/chat) renders without errors in headless Chromium.
