--- server/algo.js
@@ -1,1 +1,0 @@
-module.exports = require("./matching/algo");
