+++ src/App.js
@@ -1,15 +1,15 @@
 import React from "react";
 import { Route, Switch, Redirect } from "react-router-dom";
 import { useEffect } from "react";
-import { auth, db } from "./components/landingComp/firebase";
+import { auth, db } from "./lib/firebase";
 import store from "./store/store";
-import Loading from "./spinLoad";
+import Loading from "./components/common/SpinLoad";
 import { useSelector } from "react-redux";
 
 import MatchPage from "./pages/matchPage";
 import ChatPage from "./pages/chatPage";
 import ProfilePage from "./pages/profilePage";
-import CompleteProfile from "./components/signUpComp/completeProfile";
+import CompleteProfile from "./components/signup/completeProfile";
 
 const LandingRoutes = React.lazy(() => import("./pages/landing"));
