+++ src/components/common/spinLoad.css
