+++ src/components/landing/about.js
