+++ src/components/signup/fileInput/uploadedImg.js
