+++ src/components/signup/form.jsx
@@ -4,7 +4,7 @@ import Page1 from "./page1";
 import Page2 from "./page2";
 import Page3 from "./page3";
 import firebase from "firebase/app";
-import { db, auth } from "../landingComp/firebase";
+import { db, auth } from "../../lib/firebase";
 import store from "../../store/store";
 import { Redirect } from "react-router";
