+++ src/pages/chatPage.jsx
@@ -1,4 +1,4 @@
-import Chat from "../components/chatRoomComp/chat";
+import Chat from "../components/chat/chat";
 
 export default function () {
   return <Chat />;
