--- src/pages/loader.jsx
@@ -1,12 +1,0 @@
-import "./loaderStyles.css";
-
-export default function () {
-  return (
-    <div className="Loadcontent">
-      <div className="loading">
-        <p>Loading</p>
-        <span></span>
-      </div>
-    </div>
-  );
-}
