+++ src/pages/signUpPage.jsx
@@ -1,5 +1,5 @@
 import "./signUpStyles.css";
-import Form from "../components/signUpComp/form";
+import Form from "../components/signup/form";
 
 export default function App() {
   return (
