--- src/pages/signedIn.jsx
@@ -1,11 +1,0 @@
-import MatchPage from "./matchPage";
-import ChatPage from "./chatPage";
-import ProfilePage from "./profilePage";
-
-export default {
-  MatchPage,
-  ChatPage,
-  ProfilePage,
-};
-
-export { MatchPage, ChatPage, ProfilePage };
